Integration tests refactored into groups
