+++ tests/Integration/CmdMigrateLiveContentMigrationDataConsistencyStandardTest.php
@@ -14,11 +14,9 @@ use Newspack\ContentDiffMigrator\Tests\Integration\IntegrationTestCase;
  *
  * @group integration
  */
-class CmdMigrateLiveContentMDCSTest extends IntegrationTestCase {
+class CmdMigrateLiveContentMigrationDataConsistencyStandardTest extends IntegrationTestCase {
 	/**
-	 * "mdcs" stands for "Migration Data Consistency Standard".
-	 * 
-	 * @group mdcs
+	 * @group migration-data-consistency-standard
 	 */
 	public function test_mdcs_should_update_user_email_when_changed_on_live(): void {
 		global $wpdb;
@@ -63,9 +61,7 @@ class CmdMigrateLiveContentMDCSTest extends IntegrationTestCase {
 	}
 
 	/**
-	 * "mdcs" stands for "Migration Data Consistency Standard".
-	 * 
-	 * @group mdcs
+	 * @group migration-data-consistency-standard
 	 */
 	public function test_mdcs_should_update_user_display_name_when_changed_on_live(): void {
 		global $wpdb;
@@ -104,9 +100,7 @@ class CmdMigrateLiveContentMDCSTest extends IntegrationTestCase {
 	}
 
 	/**
-	 * "mdcs" stands for "Migration Data Consistency Standard".
-	 * 
-	 * @group mdcs
+	 * @group migration-data-consistency-standard
 	 */
 	public function test_mdcs_should_not_update_user_login_when_changed_on_live_and_should_create_new_user(): void {
 		global $wpdb;
@@ -151,9 +145,7 @@ class CmdMigrateLiveContentMDCSTest extends IntegrationTestCase {
 	}
 
 	/**
-	 * "mdcs" stands for "Migration Data Consistency Standard".
-	 * 
-	 * @group mdcs
+	 * @group migration-data-consistency-standard
 	 */
 	public function test_mdcs_should_update_attachment_caption_when_changed_on_live(): void {
 		global $wpdb;
@@ -187,9 +179,7 @@ class CmdMigrateLiveContentMDCSTest extends IntegrationTestCase {
 	}
 
 	/**
-	 * "mdcs" stands for "Migration Data Consistency Standard".
-	 * 
-	 * @group mdcs
+	 * @group migration-data-consistency-standard
 	 */
 	public function test_mdcs_should_update_attachment_description_when_changed_on_live(): void {
 		global $wpdb;
@@ -223,9 +213,7 @@ class CmdMigrateLiveContentMDCSTest extends IntegrationTestCase {
 	}
 
 	/**
-	 * "mdcs" stands for "Migration Data Consistency Standard".
-	 * 
-	 * @group mdcs
+	 * @group migration-data-consistency-standard
 	 */
 	public function test_mdcs_should_update_attachment_alt_text_when_changed_on_live(): void {
 		global $wpdb;
@@ -269,9 +257,7 @@ class CmdMigrateLiveContentMDCSTest extends IntegrationTestCase {
 	}
 
 	/**
-	 * "mdcs" stands for "Migration Data Consistency Standard".
-	 * 
-	 * @group mdcs
+	 * @group migration-data-consistency-standard
 	 */
 	public function test_mdcs_should_update_attachment_media_credit_when_changed_on_live(): void {
 		global $wpdb;
@@ -315,9 +301,7 @@ class CmdMigrateLiveContentMDCSTest extends IntegrationTestCase {
 	}
 
 	/**
-	 * "mdcs" stands for "Migration Data Consistency Standard".
-	 * 
-	 * @group mdcs
+	 * @group migration-data-consistency-standard
 	 */
 	public function test_mdcs_should_update_term_slug_when_changed_on_live(): void {
 		global $wpdb;
@@ -369,9 +353,7 @@ class CmdMigrateLiveContentMDCSTest extends IntegrationTestCase {
 	}
 
 	/**
-	 * "mdcs" stands for "Migration Data Consistency Standard".
-	 * 
-	 * @group mdcs
+	 * @group migration-data-consistency-standard
 	 */
 	public function test_mdcs_should_update_term_description_when_changed_on_live(): void {
 		global $wpdb;
@@ -421,9 +403,7 @@ class CmdMigrateLiveContentMDCSTest extends IntegrationTestCase {
 	}
 
 	/**
-	 * "mdcs" stands for "Migration Data Consistency Standard".
-	 * 
-	 * @group mdcs
+	 * @group migration-data-consistency-standard
 	 */
 	public function test_mdcs_should_not_update_term_name_when_changed_on_live(): void {
 		global $wpdb;
